:pencil: renamed ***StackConsts from ***Consts
